+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] 2020-05-28
+
+### Added
+
+- Support for receipts API
+- Edit organization's receipts settings
+- Check domain availability
+- Select organization's domain
+
 ## [2.2.3] 2020-03-05
 
 ### Fixed